Move entrypoint to main.py
