+++ asreviewcontrib/semantic_clustering/interactive.py
@@ -125,7 +125,3 @@ def run_app():
 
     # Run the application
     app.run_server(debug=True)
-
-
-if __name__ == "__main__":
-    run_app()
